# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Moskollektor Backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/sensors": {
      "get": {
        "tags": [
          "sensors"
        ],
        "summary": "Get Sensors",
        "description": "List sensors within the caller's scope.\n\nArgs:\n    facility_id: Optional exact facility filter.\n    hierarchy_node_id: Optional exact hierarchy node filter.\n    type: Optional exact sensor_type_id filter.\n    current_state: Accepted for contract completeness; no real\n        current_state data yet (see src/schemas/sensor.py).\n    risk_level: Accepted for contract completeness; same caveat.\n    query: Optional case-insensitive substring match.\n    cursor: Opaque pagination cursor.\n    limit: Page size, 1-200.\n    user: The authenticated caller, injected by get_current_user.\n\nReturns:\n    A SensorListEnvelope containing only sensors the caller may see.",
        "operationId": "get_sensors_api_v1_sensors_get",
        "parameters": [
          {
            "name": "facility_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Facility Id"
            }
          },
          {
            "name": "hierarchy_node_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hierarchy Node Id"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "name": "current_state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Current State"
            }
          },
          {
            "name": "risk_level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Risk Level"
            }
          },
          {
            "name": "query",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Query"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SensorListEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sensors/{sensor_id}": {
      "get": {
        "tags": [
          "sensors"
        ],
        "summary": "Get Sensor",
        "description": "Fetch one sensor's detail card, enforcing the caller's scope.\n\nArgs:\n    sensor_id: The requested sensor id.\n    user: The authenticated caller, injected by get_current_user.\n\nReturns:\n    The sensor's full detail card.\n\nRaises:\n    ApiError: 404 if the sensor does not exist; 403 if it exists but is\n        outside scope (orphans are out of scope for anyone restricted).",
        "operationId": "get_sensor_api_v1_sensors__sensor_id__get",
        "parameters": [
          {
            "name": "sensor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sensor Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SensorDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sensors/{sensor_id}/series": {
      "get": {
        "tags": [
          "sensor-series"
        ],
        "summary": "Get Series",
        "description": "Return a sensor's time series, in the format matching its value_type.\n\nArgs:\n    sensor_id: The requested sensor id.\n    from_: Optional inclusive start of the time range.\n    to: Optional inclusive end of the time range.\n    granularity: Optional bucket size for numeric downsampling\n        (e.g. \"5m\", \"1h\", \"1d\"); ignored for categorical sensors.\n    user: The authenticated caller, injected by get_current_user.\n\nReturns:\n    A NumericSeriesOut (points/thresholds/missing_intervals) or a\n    CategoricalSeriesOut (intervals), matching the sensor's value_type.\n\nRaises:\n    ApiError: 404 if the sensor does not exist; 403 if out of scope.",
        "operationId": "get_series_api_v1_sensors__sensor_id__series_get",
        "parameters": [
          {
            "name": "sensor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sensor Id"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "name": "granularity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Granularity"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/NumericSeriesOut"
                    },
                    {
                      "$ref": "#/components/schemas/CategoricalSeriesOut"
                    }
                  ],
                  "title": "Response Get Series Api V1 Sensors  Sensor Id  Series Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/system/scenarios": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "List Scenarios",
        "description": "List the fixture-scenario library available for controlled demonstration.",
        "operationId": "list_scenarios_api_v1_system_scenarios_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioListEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/system/scenarios/{scenario_id}/activate": {
      "post": {
        "tags": [
          "system"
        ],
        "summary": "Activate Scenario Endpoint",
        "description": "Activate a fixture scenario against a real facility.",
        "operationId": "activate_scenario_endpoint_api_v1_system_scenarios__scenario_id__activate_post",
        "parameters": [
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scenario Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivateScenarioRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivateScenarioResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ActivateScenarioRequest": {
        "properties": {
          "facility_id": {
            "type": "string",
            "title": "Facility Id"
          },
          "seed": {
            "type": "integer",
            "title": "Seed",
            "default": 42
          }
        },
        "type": "object",
        "required": [
          "facility_id"
        ],
        "title": "ActivateScenarioRequest",
        "description": "Body for POST /api/v1/system/scenarios/{scenario_id}/activate."
      },
      "ActivateScenarioResult": {
        "properties": {
          "scenario_id": {
            "type": "string",
            "title": "Scenario Id"
          },
          "facility_id": {
            "type": "string",
            "title": "Facility Id"
          },
          "readings_inserted": {
            "type": "integer",
            "title": "Readings Inserted"
          }
        },
        "type": "object",
        "required": [
          "scenario_id",
          "facility_id",
          "readings_inserted"
        ],
        "title": "ActivateScenarioResult",
        "description": "Result of a scenario activation."
      },
...
      "CategoricalSeriesOut": {
        "properties": {
          "value_type": {
            "type": "string",
            "const": "categorical",
            "title": "Value Type",
            "default": "categorical"
          },
          "intervals": {
            "items": {
              "$ref": "#/components/schemas/CategoricalInterval"
            },
            "type": "array",
            "title": "Intervals"
          }
        },
        "type": "object",
        "required": [
          "intervals"
        ],
        "title": "CategoricalSeriesOut",
        "description": "The categorical series response shape."
      },
...
      "CurrentReading": {
        "properties": {
          "value": {
            "type": "string",
            "title": "Value"
          },
          "numeric_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Numeric Value"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "measured_at": {
            "type": "string",
            "format": "date-time",
            "title": "Measured At"
          }
        },
        "type": "object",
        "required": [
          "value",
          "numeric_value",
          "unit",
          "measured_at"
        ],
        "title": "CurrentReading",
        "description": "The most recent reading for a sensor, if any exist yet."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NumericPoint": {
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "value": {
            "type": "number",
            "title": "Value"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "quality": {
            "type": "string",
            "title": "Quality"
          }
        },
        "type": "object",
        "required": [
          "timestamp",
          "value",
          "state",
          "quality"
        ],
        "title": "NumericPoint",
        "description": "One raw or bucket-averaged numeric reading."
      },
      "NumericSeriesOut": {
        "properties": {
          "value_type": {
            "type": "string",
            "const": "numeric",
            "title": "Value Type",
            "default": "numeric"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/NumericPoint"
            },
            "type": "array",
            "title": "Points"
          },
          "thresholds": {
            "items": {
              "$ref": "#/components/schemas/Threshold"
            },
            "type": "array",
            "title": "Thresholds"
          },
          "missing_intervals": {
            "items": {
              "$ref": "#/components/schemas/MissingInterval"
            },
            "type": "array",
            "title": "Missing Intervals"
          }
        },
        "type": "object",
        "required": [
          "points",
          "thresholds",
          "missing_intervals"
        ],
        "title": "NumericSeriesOut",
        "description": "The numeric series response shape."
      },
...
      "ScenarioListEnvelope": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ScenarioOut"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ScenarioListEnvelope",
        "description": "The standard {data} envelope for GET /api/v1/system/scenarios."
      },
...
      "SensorDetailOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "channel_id": {
            "type": "string",
            "title": "Channel Id"
          },
          "tag": {
            "type": "string",
            "title": "Tag"
          },
          "sensor_type": {
            "type": "string",
            "title": "Sensor Type"
          },
          "system_type": {
            "type": "string",
            "title": "System Type"
          },
          "value_type": {
            "type": "string",
            "title": "Value Type"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "facility_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Facility Id"
          },
          "hierarchy_node_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hierarchy Node Id"
          },
          "has_geolocation": {
            "type": "boolean",
            "title": "Has Geolocation",
            "default": false
          },
          "position": {
            "type": "null",
            "title": "Position"
          },
          "hierarchy_path": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Hierarchy Path"
          },
          "current_reading": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CurrentReading"
              },
              {
                "type": "null"
              }
            ]
          },
          "current_state": {
            "type": "string",
            "title": "Current State"
          },
          "forecast_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forecast Summary"
          },
          "data_health": {
            "type": "string",
            "title": "Data Health"
          },
          "maintenance_state": {
            "type": "string",
            "title": "Maintenance State"
          }
        },
        "type": "object",
        "required": [
          "id",
          "channel_id",
          "tag",
          "sensor_type",
          "system_type",
          "value_type",
          "display_name",
          "facility_id",
          "hierarchy_node_id",
          "hierarchy_path",
          "current_reading",
          "current_state",
          "forecast_summary",
          "data_health",
          "maintenance_state"
        ],
        "title": "SensorDetailOut",
        "description": "The full sensor detail card (superset of the list item shape)."
      },
      "SensorListEnvelope": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/SensorOut"
            },
            "type": "array",
            "title": "Data"
          },
          "meta": {
            "$ref": "#/components/schemas/SensorListMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "SensorListEnvelope",
        "description": "The standard {data, meta} envelope for GET /api/v1/sensors."
      },
      "SensorListMeta": {
        "properties": {
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          }
        },
        "type": "object",
        "required": [
          "next_cursor",
          "total",
          "generated_at"
        ],
        "title": "SensorListMeta",
        "description": "Cursor-pagination metadata for the sensor list envelope."
      },
      "SensorOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "channel_id": {
            "type": "string",
            "title": "Channel Id"
          },
          "tag": {
            "type": "string",
            "title": "Tag"
          },
          "sensor_type": {
            "type": "string",
            "title": "Sensor Type"
          },
          "system_type": {
            "type": "string",
            "title": "System Type"
          },
          "value_type": {
            "type": "string",
            "title": "Value Type"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "facility_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Facility Id"
          },
          "hierarchy_node_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hierarchy Node Id"
          },
          "has_geolocation": {
            "type": "boolean",
            "title": "Has Geolocation",
            "default": false
          },
          "position": {
            "type": "null",
            "title": "Position"
          }
        },
        "type": "object",
        "required": [
          "id",
          "channel_id",
          "tag",
          "sensor_type",
          "system_type",
          "value_type",
          "display_name",
          "facility_id",
          "hierarchy_node_id"
        ],
        "title": "SensorOut",
        "description": "A sensor channel as it appears in the list."
      },
...
      "Threshold": {
        "properties": {
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "value": {
            "type": "number",
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "value"
        ],
        "title": "Threshold",
        "description": "A named threshold value (always empty for MVP -- see module docstring)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}